refactor(provider): extract agent-building wiring into build.rs (dirge-4y4l stage 7)
Move the DI/construction layer out of provider/mod.rs into a new
provider::build submodule: create_client, build_agent (the
variant-dispatch build_inner! macro + escalation/critic/review/
context-depth/bg-store/max-turns wiring), and the standalone builders
build_escalation_stream_fn, build_critic_fn, build_approval_fn,
build_review_stream_fn. mod.rs re-exports via 'pub use build::*' so
provider::build_agent / create_client / build_approval_fn paths are
unchanged.

build_agent read AnyAgent's private loop_tools field directly; added a
pub(crate) loop_tools() accessor (next to cache()) so the sibling
module uses the public surface. Pruned the now-unused mod.rs imports;
mod_tests.rs imports ProviderEntry directly (previously via use
super::*).

Behavior-preserving. mod.rs 2217 (pre-refactor) -> 951; build.rs 537.
Builds clean -D warnings on default + windows-default; 2374 tests
green.
